Add files via upload
In this file I try to calculate the health care access for multiple travel times. However, the travel times are calculated with an OSM license and it seems to have API limitation. If I try to calculate it, the travel times return None at some point without a reason. If a subset of the first 5 rows is chosen, the travel times do not return None until 3000 seconds, while if we only choose one row None is only returned after 3600 seconds.